backend/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.1
flask-marshmallow==0.15.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pkgutil-resolve-name==1.3.10
pymssql==2.2.7
pyodbc==4.0.39
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.11
typing-extensions==4.5.0
tzdata==2023.3
Werkzeug==2.3.2
zipp==3.15.0
